refactor: test 패키지 정리 및 GameTable 정리
