+++ src/main/java/blackjack/domain/Game.java
@@ -8,7 +8,7 @@ import blackjack.domain.gamer.Player;
 import blackjack.domain.gamer.Players;
 import blackjack.domain.gametable.GameTable;
 import blackjack.dto.Results;
-import blackjack.utils.RandomCardDeck;
+import blackjack.domain.utils.RandomCardDeck;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -16,12 +16,13 @@ import java.util.List;
 
 public class Game {
     private final GameTable gameTable;
+    private final Players players;
     private final Participant dealer;
 
     public Game(String namesValue) {
         List<Player> playersValue = getPlayerList(namesValue);
         this.dealer = new Dealer(new Cards(Collections.emptyList()));
-        final Players players = new Players(playersValue);
+        this.players = new Players(playersValue);
 
         gameTable = new GameTable(dealer, players, new RandomCardDeck());
     }
@@ -36,11 +37,11 @@ public class Game {
     }
 
     public Results getParticipants() {
-        return gameTable.getParticipants();
+        return new Results(players, dealer);
     }
 
     public List<Player> getPlayers() {
-        return gameTable.getPlayers();
+        return players.getUnmodifiableList();
     }
 
     public Player turnForPlayer(Player player) {
