+++ src/main/java/blackjack/domain/gametable/Playable.java
@@ -1,8 +1,6 @@
 package blackjack.domain.gametable;
 
-import blackjack.domain.card.Card;
 import blackjack.domain.gamer.Participant;
-import java.util.List;
 
 public interface Playable {
     void giveCard(Participant participant);
