+++ src/main/java/blackjack/domain/utils/CardDeck.java
@@ -1,11 +1,9 @@
-package blackjack.utils;
+package blackjack.domain.utils;
 
 import blackjack.domain.card.Card;
 
 import blackjack.domain.card.Suits;
 import java.util.LinkedList;
-import java.util.List;
-import java.util.Queue;
 
 public interface CardDeck {
     void assembleWithDenominations(LinkedList<Card> cardsValue, Suits suit);
