+++ src/test/java/blackjack/domain/OutcomeTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public class OutcomeTest {
     /*
