+++ src/test/java/blackjack/domain/gamer/NameTest.java
@@ -1,0 +1,4 @@
+package blackjack.domain.gamer;
+
+public class NameTest {
+}
