+++ src/test/java/blackjack/domain/gamer/PlayersTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
