+++ src/test/java/blackjack/domain/utils/FixedCardDeck.java
@@ -1,4 +1,4 @@
-package blackjack.utils;
+package blackjack.domain.utils;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Denominations;
@@ -6,22 +6,23 @@ import blackjack.domain.card.Suits;
 
 import java.util.*;
 
-public class RandomCardDeck implements CardDeck {
+public class FixedCardDeck implements CardDeck {
     private final Queue<Card> cards;
 
-    public RandomCardDeck() {
+    public FixedCardDeck() {
         LinkedList<Card> cardsValue = new LinkedList<>();
+
         for (Suits suit : Suits.values()) {
             assembleWithDenominations(cardsValue, suit);
         }
-        Collections.shuffle(cardsValue);
+
         this.cards = cardsValue;
     }
 
     @Override
     public void assembleWithDenominations(LinkedList<Card> cardsValue, Suits suit) {
         for (Denominations denomination : Denominations.values()) {
-            cardsValue.add(Card.from(denomination, suit));
+            cardsValue.offer(Card.from(denomination, suit));
         }
     }
